# openapi.yaml
openapi: 3.0.3

info:
  title: Git-Prasaaran API
  description: Git-backed Markdown documentation publishing API
  version: 1.0.0

servers:
  - url: http://localhost:8080

tags:
  - name: Authentication
    description: Account registration and authentication
  - name: Documents
    description: Published documentation
  - name: Private Documents
    description: Authenticated user-owned private Markdown documents
  - name: Repositories
    description: Application repository contexts and their published documents
  - name: Webhooks
    description: GitHub webhook integration
  - name: Activity
    description: Persisted GitHub webhook activity
  - name: Assistant
    description: Repository documentation Q&A
  - name: Health
    description: Application health

paths:
  /api/v1/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a customer account
      operationId: registerAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationRequest"
      responses:
        "201":
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationResponse"
        "400":
          description: Invalid registration details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: An account with this email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: Sign in and obtain a JWT
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Invalid credentials

  /api/v1/documents:
    get:
      tags:
        - Documents
      summary: List published documents
      operationId: getDocuments
      responses:
        "200":
          description: All available documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"

  /api/v1/documents/{slug}:
    get:
      tags:
        - Documents
      summary: Get a published document
      operationId: getDocument
      parameters:
        - name: slug
          in: path
          required: true
          description: Document slug
          schema:
            type: string
      responses:
        "200":
          description: The document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/private-documents:
    get:
      tags:
        - Private Documents
      summary: List the authenticated user's private documents
      description: Returns only documents owned by the authenticated user.
      operationId: getPrivateDocuments
      security:
        - BearerAuth: []
      responses:
        "200":
          description: The authenticated user's private documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PrivateDocument"
        "401":
          description: Authentication is required
    post:
      tags:
        - Private Documents
      summary: Create a private document
      description: The owner is derived from the authenticated user; any supplied owner ID is ignored.
      operationId: createPrivateDocument
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrivateDocumentRequest"
      responses:
        "201":
          description: Private document created
          headers:
            Location:
              description: URI of the created private document
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivateDocument"
        "400":
          description: Invalid title, slug, or Markdown content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication is required
        "409":
          description: The owner already has a document with this slug
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /api/v1/private-documents/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Private document ID
        schema:
          type: integer
          format: int64
          minimum: 1
    get:
      tags:
        - Private Documents
      summary: Get an owned private document
      description: Documents not owned by the authenticated user are returned as not found.
      operationId: getPrivateDocument
      security:
        - BearerAuth: []
      responses:
        "200":
          description: The owned private document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivateDocument"
        "401":
          description: Authentication is required
        "404":
          description: Private document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      tags:
        - Private Documents
      summary: Update an owned private document
      description: Ownership is verified against the authenticated user; documents outside the user's scope are returned as not found.
      operationId: updatePrivateDocument
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrivateDocumentRequest"
      responses:
        "200":
          description: Private document updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivateDocument"
        "400":
          description: Invalid title, slug, or Markdown content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication is required
        "404":
          description: Private document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: The owner already has a document with this slug
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      tags:
        - Private Documents
      summary: Delete an owned private document
      description: Documents not owned by the authenticated user are returned as not found.
      operationId: deletePrivateDocument
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Private document deleted
        "401":
          description: Authentication is required
        "404":
          description: Private document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /api/v1/repositories:
    get:
      tags:
        - Repositories
      summary: List configured repositories
      description: Requires an authenticated user. Repository configuration responses do not include GitHub credentials.
      operationId: getRepositories
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Configured repositories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Repository"
        "401":
          description: Authentication is required
    post:
      tags:
        - Repositories
      summary: Register a repository
      description: Requires an authenticated user with the ADMIN role. The repository ID is generated by the server.
      operationId: createRepository
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepositoryCreateRequest"
      responses:
        "201":
          description: Repository registered
          headers:
            Location:
              description: URI of the created repository
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
        "400":
          description: Request fields are missing, blank, or exceed their maximum length
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication is required
        "403":
          description: The authenticated user does not have the ADMIN role
        "409":
          description: The GitHub owner and repository name are already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /api/v1/repositories/{repositoryId}:
    parameters:
      - name: repositoryId
        in: path
        required: true
        description: Internal application repository ID
        schema:
          type: integer
          format: int64
    get:
      tags:
        - Repositories
      summary: Get a configured repository
      description: Requires an authenticated user. Repository configuration responses do not include GitHub credentials.
      operationId: getRepository
      security:
        - BearerAuth: []
      responses:
        "200":
          description: The repository configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
        "401":
          description: Authentication is required
        "404":
          description: Repository not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      tags:
        - Repositories
      summary: Update repository configuration or active state
      description: Requires an authenticated user with the ADMIN role. Set active to false to deactivate; repository deletion is not supported.
      operationId: updateRepository
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepositoryPatchRequest"
      responses:
        "200":
          description: The updated repository configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
        "400":
          description: No update fields were provided, or a supplied string is blank or exceeds its maximum length
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication is required
        "403":
          description: The authenticated user does not have the ADMIN role
        "404":
          description: Repository not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: The updated GitHub owner and repository name are already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /api/v1/repositories/{repositoryId}/documents:
    parameters:
      - name: repositoryId
        in: path
        required: true
        description: Internal application repository ID
        schema:
          type: integer
          format: int64
    get:
      tags:
        - Repositories
        - Documents
      summary: List documents from a configured repository
      description: Requires an authenticated user. Documents are retrieved using the selected repository configuration. An inactive repository or retrieval failure currently returns 502.
      operationId: getRepositoryDocuments
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Documents available from the selected repository
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
        "401":
          description: Authentication is required
        "404":
          description: Repository not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: The repository is inactive or document retrieval failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /api/v1/repositories/{repositoryId}/documents/{slug}:
    parameters:
      - name: repositoryId
        in: path
        required: true
        description: Internal application repository ID
        schema:
          type: integer
          format: int64
      - name: slug
        in: path
        required: true
        description: Document slug within the selected repository
        schema:
          type: string
    get:
      tags:
        - Repositories
        - Documents
      summary: Get a document from a configured repository
      description: Requires an authenticated user. A repository-specific Redis cache is used; cache keys include the internal repository ID. An inactive repository or retrieval failure currently returns 502.
      operationId: getRepositoryDocument
      security:
        - BearerAuth: []
      responses:
        "200":
          description: The document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "401":
          description: Authentication is required
        "404":
          description: Repository or document not found. A missing document returns an empty response body; an unknown repository returns an ApiError.
        "502":
          description: The repository is inactive or document retrieval failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /api/v1/activity:
    get:
      tags:
        - Activity
      summary: List webhook activity
      description: Requires an authenticated user. Results are ordered newest first; historical events may have a null repositoryId.
      operationId: getActivity
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ActivityPage"
        - $ref: "#/components/parameters/ActivitySize"
        - $ref: "#/components/parameters/ActivityStatus"
        - $ref: "#/components/parameters/ActivityEventType"
        - $ref: "#/components/parameters/ActivityFrom"
        - $ref: "#/components/parameters/ActivityTo"
      responses:
        "200":
          description: A page of webhook events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityPage"
        "400":
          description: Invalid pagination, filter, or date range
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication is required

  /api/v1/repositories/{repositoryId}/activity:
    get:
      tags:
        - Activity
        - Repositories
      summary: List activity for a repository
      description: Requires an authenticated user. Only webhook events associated with this repository are returned.
      operationId: getRepositoryActivity
      security:
        - BearerAuth: []
      parameters:
        - name: repositoryId
          in: path
          required: true
          description: Internal application repository ID
          schema:
            type: integer
            format: int64
        - $ref: "#/components/parameters/ActivityPage"
        - $ref: "#/components/parameters/ActivitySize"
        - $ref: "#/components/parameters/ActivityStatus"
        - $ref: "#/components/parameters/ActivityEventType"
        - $ref: "#/components/parameters/ActivityFrom"
        - $ref: "#/components/parameters/ActivityTo"
      responses:
        "200":
          description: A page of webhook events associated with the repository
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityPage"
        "400":
          description: Invalid pagination, filter, or date range
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication is required
        "404":
          description: Repository not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /api/v1/webhooks/github:
    post:
      tags:
        - Webhooks
      summary: Process a GitHub webhook
      operationId: processGithubWebhook
      parameters:
        - name: X-Hub-Signature-256
          in: header
          required: true
          description: GitHub HMAC SHA-256 signature
          schema:
            type: string
            example: sha256=abc123
        - name: X-GitHub-Event
          in: header
          required: true
          description: GitHub event type
          schema:
            type: string
            example: push
        - name: X-GitHub-Delivery
          in: header
          required: true
          description: Unique GitHub delivery identifier
          schema:
            type: string
            example: 72d3162e-cc78-11e3-81ab-4c9367b1baf2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "202":
          description: Push delivery accepted; unsupported event types return 200 with status ignored
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - event
                  - duplicate
                  - processed
                properties:
                  status:
                    type: string
                    example: accepted
                  event:
                    type: string
                    example: push
                  duplicate:
                    type: boolean
                  processed:
                    type: boolean
        "200":
          description: Unsupported event ignored
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - event
                properties:
                  status:
                    type: string
                    example: ignored
                  event:
                    type: string
        "400":
          description: Invalid webhook request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Invalid webhook signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /api/v1/health:
    get:
      tags:
        - Health
      summary: Check application health
      operationId: health
      responses:
        "200":
          description: Application is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"


  /api/v1/assistant/ask:
    post:
      tags:
        - Assistant
      summary: Ask a question about a registered repository's Markdown documentation
      description: Requires authentication. The answer is generated from bounded documentation context for the selected active repository.
      operationId: askDocumentationAssistant
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssistantRequest'
      responses:
        '200':
          description: Answer generated from repository documentation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssistantResponse'
        '400':
          description: Invalid request
        '401':
          description: Authentication required
        '409':
          description: Repository is inactive
        '422':
          description: No usable documentation found
        '503':
          description: AI provider unavailable

components:
  parameters:
    ActivityPage:
      name: page
      in: query
      required: false
      description: Zero-based page number
      schema:
        type: integer
        minimum: 0
        default: 0
    ActivitySize:
      name: size
      in: query
      required: false
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ActivityStatus:
      name: status
      in: query
      required: false
      schema:
        type: string
        enum: [RECEIVED, PROCESSED, FAILED]
    ActivityEventType:
      name: eventType
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 100
    ActivityFrom:
      name: from
      in: query
      required: false
      description: Inclusive lower bound for createdAt
      schema:
        type: string
        format: date-time
    ActivityTo:
      name: to
      in: query
      required: false
      description: Inclusive upper bound for createdAt
      schema:
        type: string
        format: date-time
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegistrationRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          example: customer@example.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          description: Account password; stored as a password hash by the server

    RegistrationResponse:
      type: object
      required:
        - id
        - email
      properties:
        id:
          type: integer
          format: int64
          example: 42
        email:
          type: string
          format: email
          example: customer@example.com

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: customer@example.com
        password:
          type: string
          format: password

    LoginResponse:
      type: object
      required:
        - accessToken
        - tokenType
        - expiresIn
      properties:
        accessToken:
          type: string
          description: JWT access token
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
          description: Token lifetime in seconds
          example: 900

    AssistantRequest:
      type: object
      required:
        - repositoryId
        - question
      properties:
        repositoryId:
          type: integer
          format: int64
          minimum: 1
        question:
          type: string
          minLength: 1
          maxLength: 2000
    AssistantResponse:
      type: object
      required:
        - answer
        - sources
        - model
      properties:
        answer:
          type: string
        sources:
          type: array
          items:
            type: string
        model:
          type: string
    Document:
      type: object
      required:
        - slug
        - title
        - content
        - html
        - sourcePath
      properties:
        slug:
          type: string
          example: getting-started
        title:
          type: string
          example: Getting Started
        description:
          type: string
          nullable: true
          example: Introduction to Git-Prasaaran
        content:
          type: string
          description: Raw Markdown source
        html:
          type: string
          description: Sanitized HTML rendered by the backend
        sourcePath:
          type: string
          example: docs/architecture.md

    PrivateDocumentRequest:
      type: object
      required:
        - title
        - slug
        - content
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          type: string
          minLength: 1
          maxLength: 120
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
          description: Lowercase letters and numbers separated by single hyphens
        content:
          type: string
          minLength: 1
          maxLength: 100000
          description: Markdown source; stored privately in PostgreSQL

    PrivateDocument:
      type: object
      required:
        - id
        - title
        - slug
        - content
        - html
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          maxLength: 200
        slug:
          type: string
          maxLength: 120
        content:
          type: string
          description: Private Markdown source
        html:
          type: string
          description: Sanitized rendered Markdown HTML
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: DOCUMENT_NOT_FOUND
        message:
          type: string
          example: Document not found

    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: REPOSITORY_NOT_FOUND
        message:
          type: string
          example: Repository not found

    ActivityPage:
      type: object
      required:
        - items
        - page
        - size
        - totalElements
        - totalPages
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ActivityEvent"
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0

    ActivityEvent:
      type: object
      required:
        - id
        - deliveryId
        - eventType
        - commitSha
        - status
        - createdAt
        - processedAt
        - repositoryId
      properties:
        id:
          type: integer
          format: int64
        deliveryId:
          type: string
        eventType:
          type: string
        commitSha:
          type: string
          nullable: true
        status:
          type: string
          enum: [RECEIVED, PROCESSED, FAILED]
        createdAt:
          type: string
          format: date-time
        processedAt:
          type: string
          format: date-time
          nullable: true
        repositoryId:
          type: integer
          format: int64
          nullable: true

    RepositoryCreateRequest:
      type: object
      required:
        - owner
        - name
        - branch
        - contentPath
      properties:
        owner:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '.*\S.*'
          description: GitHub repository owner; must not be blank
        name:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '.*\S.*'
          description: GitHub repository name; must not be blank
        branch:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '.*\S.*'
          description: Git branch; must not be blank
        contentPath:
          type: string
          minLength: 1
          maxLength: 500
          pattern: '.*\S.*'
          description: Path containing documentation content; must not be blank

    RepositoryPatchRequest:
      type: object
      description: At least one non-null property must be supplied. Null properties are treated as omitted.
      anyOf:
        - required:
            - owner
          properties:
            owner:
              type: string
        - required:
            - name
          properties:
            name:
              type: string
        - required:
            - branch
          properties:
            branch:
              type: string
        - required:
            - contentPath
          properties:
            contentPath:
              type: string
        - required:
            - active
          properties:
            active:
              type: boolean
      properties:
        owner:
          type: string
          nullable: true
          minLength: 1
          maxLength: 255
          pattern: '.*\S.*'
        name:
          type: string
          nullable: true
          minLength: 1
          maxLength: 255
          pattern: '.*\S.*'
        branch:
          type: string
          nullable: true
          minLength: 1
          maxLength: 255
          pattern: '.*\S.*'
        contentPath:
          type: string
          nullable: true
          minLength: 1
          maxLength: 500
          pattern: '.*\S.*'
        active:
          type: boolean
          nullable: true
          description: Set to false to deactivate or true to activate the repository

    Repository:
      type: object
      description: Application-level configuration referencing a GitHub repository. No credentials or tokens are returned.
      required:
        - id
        - owner
        - name
        - branch
        - contentPath
        - active
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: Stable internal repository identifier
        owner:
          type: string
        name:
          type: string
        branch:
          type: string
        contentPath:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: UP